lesson_8_level_5

+++ src/main/java/student_einars_marhilevics/lesson_8/level_4/Square.java
@@ -1,11 +1,11 @@
 package student_einars_marhilevics.lesson_8.level_4;
 
 public class Square extends Shape{
-    int side;
+    double side;
 
     public Square(String title, double side) {
         super(title);
-        this.side = (int) side;
+        this.side = side;
     }
 
     @Override
